+++ src/datumaro/plugins/data_formats/ade20k2017.py
@@ -12,7 +12,7 @@ from typing import List, Optional
 
 import numpy as np
 
-from datumaro.components.annotation import AnnotationType, CompiledMask, LabelCategories, Mask
+from datumaro.components.annotation import AnnotationType, ExtractedMask, LabelCategories
 from datumaro.components.dataset_base import DatasetBase, DatasetItem
 from datumaro.components.errors import InvalidAnnotationError
 from datumaro.components.format_detection import FormatDetectionContext
@@ -97,7 +97,6 @@ class Ade20k2017Base(DatasetBase):
                     continue
 
                 mask = lazy_image(mask_path, loader=self._load_instance_mask)
-                mask = CompiledMask(instance_mask=mask)
 
                 for v in item_info:
                     if v["part_level"] != part_level:
@@ -108,9 +107,10 @@ class Ade20k2017Base(DatasetBase):
                     attributes = {k: True for k in v["attributes"]}
 
                     item_annotations.append(
-                        Mask(
+                        ExtractedMask(
+                            index_mask=mask,
+                            index=instance_id,
                             label=label_id,
-                            image=mask.lazy_extract(instance_id),
                             id=instance_id,
                             attributes=attributes,
                             z_order=part_level,