+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/1442>)
 - Get target information for tabular dataset
   (<https://github.com/openvinotoolkit/datumaro/pull/1471>)
+- Add ExtractedMask and update importers who can use it to use it
+  (<https://github.com/openvinotoolkit/datumaro/pull/1480>)
 
 ## May 2024 Release 1.6.1
 ### Enhancements